Fusion
First, we separate Card creation and drawing.
Then, we make openedStates Transform object that emits
signal with updated values when any card changes opened
state.
After that, we make openedCount object that emits the number
of opened cards only.
Our game logic is to close cards when we already opened
enough to form a match pair.
We can achieve this by drawing invisible overlay that will
handle click to close cards.
To actually subscribe on these signals and call required
function, we wrap our game form into TConstruct. It manages
our subscriptions while the form is shown, unsubscribing
when the form is destroyed.
* Never forget to unsubscribe when you're done *